cleanup readme, use `toml-test` via a submodule instead of cloning
