+++ bin/test.sh
