+++ tests/toml_test.rs
@@ -1,45 +1,24 @@
-//! Runs TOML's official test suite: https://github.com/toml-lang/toml-test
-//!
-//! To run: `cargo t toml_test -- --nocapture`
-//!
-//! Assumes git is installed and works.
-
 use {
 	boml::{
 		prelude::*,
 		types::{TomlDate, TomlOffset, TomlTime},
 	},
 	json::JsonValue,
-	std::{env, fs, process::Command},
+	std::{env, fs},
 };
 
+/// Run BOML against the official TOML test suite. This test fails if any valid
+/// TOML tests fail. It ignores any invalid TOML tests that pass, since these
+/// are currently quite pedantic and I'm not too worried about passing them.
+///
+/// To see what TOML tests passed/failed, run this test without capturing its
+/// `stdout`:
+/// ```
+/// cargo t toml_test -- --nocapture
+/// ```
 #[test]
 fn toml_test() {
-	// Gets us into boml/target/toml-test/tests
-	// the toml-test directory is the cloned toml-test repo
-	let mut cwd = env::current_dir().unwrap().join("target");
-	if !cwd.exists() {
-		fs::create_dir(&cwd).unwrap();
-	}
-	cwd.push("toml-test");
-	if !cwd.exists() {
-		let git = Command::new("git")
-			.arg("clone")
-			.arg("https://github.com/toml-lang/toml-test")
-			.current_dir(cwd.parent().unwrap())
-			.status();
-
-		if git.is_err() || !git.unwrap().success() {
-			panic!("Git failed to clone the test suite");
-		}
-	}
-	cwd.push("tests");
-
-	env::set_current_dir(&cwd).unwrap();
-	let git = Command::new("git").arg("pull").status();
-	if git.is_err() || !git.unwrap().success() {
-		println!("[WARN] Git failed to pull the test suite - tests may be out of date");
-	}
+	enter_toml_test_folder();
 
 	let files = fs::read_to_string("./files-toml-1.1.0").unwrap();
 	let mut lines = files.lines().peekable();
@@ -53,17 +32,15 @@ fn toml_test() {
 
 	// Invalid TOML tests
 	while let Some(file) = lines.next() {
-		println!("Testing `{file}`");
-
 		let Ok(input) = fs::read_to_string(file) else {
-			println!("WARNING: Failed to read test, skipping");
+			println!("WARNING: Failed to read test `{file}`, skipping");
 			tests_failed_to_read += 1;
 			continue;
 		};
 		let toml = boml::parse(&input);
 
 		if toml.is_ok() {
-			println!("WARNING: Invalid test succeeded");
+			println!("WARNING: Invalid test `{file}` succeeded");
 			invalid_tests_passed += 1;
 		} else {
 			invalid_tests_failed += 1;
@@ -85,7 +62,6 @@ fn toml_test() {
 	// Valid TOML tests
 	while let Some(expectation_file) = lines.next() {
 		let file = lines.next().unwrap();
-		println!("Testing `{file}`");
 
 		let expected_response = fs::read_to_string(expectation_file).unwrap();
 		let input = fs::read_to_string(file).unwrap();
@@ -93,7 +69,7 @@ fn toml_test() {
 		let toml = match boml::parse(&input) {
 			Ok(toml) => toml,
 			Err(err) => {
-				println!("WARNING: Test failed: {err:?}");
+				println!("ERROR: Valid test `{file}` failed: {err:?}");
 				valid_tests_failed += 1;
 				continue;
 			}
@@ -105,7 +81,7 @@ fn toml_test() {
 		if json_equals_toml(&expected_response, &val, file) {
 			valid_tests_passed += 1;
 		} else {
-			println!("WARNING: JSON != TOML:\n{expected_response}\n//\n{val:#?}");
+			println!("ERROR: JSON != TOML:\n{expected_response}\n//\n{val:#?}");
 			valid_tests_failed += 1;
 		}
 	}
@@ -128,6 +104,70 @@ fn toml_test() {
 	}
 }
 
+/// Get a rough estimate of boml's performance by parsing the entire TOML test
+/// suite. Note that this time gets affected by lots of other things (e.g.
+/// reading files from disk), so it's just a ballpark performance measure.
+///
+/// Run this test with:
+/// ```
+/// cargo +nightly t toml_test_speed -- -Zunstable-options --report-time --include-ignored
+/// ```
+///
+/// You can find out how many lines of code are in the test suite with tokei,
+/// e.g. in the `toml-test` folder:
+/// ```
+/// cat files-toml-1.1.0 | grep '.toml' | xargs tokei
+/// ```
+#[ignore]
+#[test]
+fn toml_test_speed() {
+	let cwd = env::current_dir()
+		.unwrap()
+		.join("target")
+		.join("toml-test")
+		.join("tests");
+	env::set_current_dir(&cwd).unwrap();
+
+	let files = fs::read_to_string("./files-toml-1.1.0").unwrap();
+	let mut lines = files.lines().peekable();
+
+	// Invalid TOML tests
+	while let Some(file) = lines.next() {
+		let Ok(input) = fs::read_to_string(file) else {
+			continue;
+		};
+		let _toml = boml::parse(&input);
+
+		if !lines.peek().unwrap().contains("invalid") {
+			break;
+		}
+	}
+
+	// Valid TOML tests
+	while let Some(_expectation_file) = lines.next() {
+		let file = lines.next().unwrap();
+		let input = fs::read_to_string(file).unwrap();
+		let _toml = boml::parse(&input);
+	}
+}
+
+/// Ensure the `toml-test` test suite is downloaded, then set our current
+/// directory to that folder so we can run its tests.
+fn enter_toml_test_folder() {
+	let toml_test_folder = env::current_dir().unwrap().join("toml-test");
+	if !toml_test_folder.exists() {
+		panic!(
+			"You need to update the `toml-test` git submodule so boml's tests can access it. You can do this with:\n\tgit submodule update --init\n\nAfter running that command, you'll see a new `toml-test` directory, which has TOML's official test suite. You can then rerun boml's tests and boml will run the official test suite."
+		)
+	}
+
+	env::set_current_dir(toml_test_folder.join("tests")).unwrap();
+}
+
+/// The TOML test suite works by having a TOML file and then the same data
+/// encoded in JSON. To pass the valid tests, you compare what you parsed from
+/// the TOML file to what an official JSON parser parsed. This function handles
+/// that by comparing `boml` to the `json` crate.
 fn json_equals_toml(json: &JsonValue, toml: &TomlValue, test_file: &str) -> bool {
 	if json.is_object() {
 		if json.has_key("type") && json.has_key("value") {
